+++ common/src/filters/mod.rs
@@ -1,5 +1,4 @@
 pub mod angle_pid;
-pub mod linear_lut;
 pub mod motor_lin;
 pub mod rate_pid;
 pub mod statistics;
