--- common/src/parsers/crsf/address.rs
@@ -1,1 +1,0 @@
-
