+++ common/src/signals/mod.rs
@@ -1,13 +1,11 @@
-use core::sync::atomic::{AtomicBool, AtomicU16, AtomicUsize};
+use core::sync::atomic::{AtomicBool, AtomicU16};
 
-use embassy_time::Duration;
 use nalgebra::UnitQuaternion;
 
 use crate::{
     IMU_COUNT, NUM_MAG,
     calibration::Calibrate,
     errors::HolsatusError,
-    health::redundancy::Mode,
     sync::{broadcast::Broadcast, channel::Channel},
     tasks::{calibrator::CalibratorState, eskf::EskfEstimate},
     types::{
@@ -54,13 +52,11 @@ pub fn register_error(error: impl Into<HolsatusError>) {
     ERR_CHANNEL.send_immediate(error.into());
 }
 
-// Raw sensor data. These are not calibrated in any way!
-// NOTE: For redundant sensors we need an array of these!
+/// Raw sensor data
 pub static RAW_IMU_DATA: Watch<Imu6DofData<f32>> = Watch::new();
 pub static RAW_MAG_DATA: Watch<Imu9DofData<f32>> = Watch::new();
 
-// Calibrated sensor data.
-// NOTE: For redundant sensors we need an array of these!
+/// Calibrated sensor data.
 pub static CAL_IMU_DATA: Watch<Imu6DofData<f32>> = Watch::new();
 pub static CAL_MAG_DATA: Watch<Imu9DofData<f32>> = Watch::new();
 
@@ -83,9 +79,6 @@ pub static RC_CHANNELS_RAW: Watch<Option<[u16; 16]>> = Watch::new();
 pub static RC_ANALOG_UNIT: Watch<RcAnalog> = Watch::new();
 pub static RC_STATUS: Watch<RcStatus> = Watch::new();
 
-pub static RATE_PIPELINE_TIME: Watch<[Duration; 4]> = Watch::new();
-
-//
 pub static RATE_PID_TERMS: Watch<[PidTerms; 3]> = Watch::new();
 pub static COMP_FUSE_GYR: Watch<[f32; 3]> = Watch::new();
 
@@ -98,9 +91,6 @@ pub static ATTITUDE_INT_EN: Watch<bool> = Watch::new();
 pub static ARMING_BLOCKER: Watch<ArmingBlocker> = Watch::new();
 
 pub static MOTORS_STATE: Watch<MotorsState> = Watch::new();
-
-// Motor speeds published by the rate controller
-pub static CTRL_MOTORS: Watch<[f32; 4]> = Watch::new();
 
 /// Whether the USB is connected or not
 pub static USB_CONNECTED: Watch<bool> = Watch::new();
@@ -116,7 +106,6 @@ pub static BLACKBOX_QUEUE: Channel<LoggableType, 10> = Channel::new();
 pub static OUTPUT_OVERRIDE: AtomicBool = AtomicBool::new(false);
 pub static IN_FLIGHT: AtomicBool = AtomicBool::new(false);
 pub static CONTROL_FREQUENCY: AtomicU16 = AtomicU16::new(0);
-pub static ACTIVE_IMU: AtomicUsize = AtomicUsize::new(0);
 
 #[macro_export]
 macro_rules! get_ctrl_freq {
@@ -140,8 +129,6 @@ multi_watch!(CAL_MULTI_IMU_DATA, Imu6DofData<f32>, IMU_COUNT, 2);
 
 multi_watch!(RAW_MULTI_MAG_DATA, [f32; 3], NUM_MAG, 2);
 multi_watch!(CAL_MULTI_MAG_DATA, [f32; 3], NUM_MAG, 2);
-
-pub static IMU_MODES: Watch<[Mode; IMU_COUNT]> = Watch::new();
 
 #[cfg(feature = "mavlink")]
 pub use crate::mavlink::Message;
