+++ common/src/sync/mod.rs
@@ -1,5 +1,4 @@
 pub mod broadcast;
-pub mod cancel;
 pub mod channel;
 pub mod lazy_lock;
 pub mod once_lock;
