+++ common/src/types/mod.rs
@@ -3,8 +3,6 @@ pub mod blackbox;
 pub mod config;
 pub mod control;
 pub mod device;
-pub mod flight_mode;
-// pub mod error;
-pub mod gcs_comm;
+// pub mod flight_mode;
 pub mod measurements;
 pub mod status;
