+++ common/src/utils/mod.rs
@@ -3,6 +3,5 @@ pub mod bytes_reader;
 pub mod const_default;
 pub mod crc8;
 pub mod func;
-pub mod matrix;
 pub mod rot_matrix;
 pub mod u_types;
