+++ device/stm32f405-dev/src/main.rs
@@ -80,9 +80,6 @@ async fn main(level_t_spawner: embassy_executor::Spawner) {
     level_t_spawner.spawn(common::tasks::arm_blocker::main().unwrap());
     level_t_spawner.spawn(common::tasks::eskf::main().unwrap());
 
-    #[cfg(feature = "mpc")]
-    level_t_spawner.spawn(common::tasks::controller_mpc::main().unwrap());
-
     level_t_spawner.spawn(common::tasks::in_flight_estimator::main().unwrap());
 
     // -------------------------- fin ---------------------------
